:building_construction: move front-end login code
